refactor ImportSpotifyPlaylistsActivity.kt & SpotifyReceiveActivity.kt into /external/spotify
