+++ app/src/androidTest/java/com/github/fribourgsdp/radio/ImportSpotifyPlaylistsActivityTest.kt
@@ -1,6 +1,8 @@
 package com.github.fribourgsdp.radio
 
 import com.github.fribourgsdp.radio.data.Song
+import com.github.fribourgsdp.radio.external.spotify.ImportSpotifyPlaylistsActivity
+import com.github.fribourgsdp.radio.external.spotify.SPOTIFY_GET_PLAYLIST_IDS_BASE_URL
 import com.github.fribourgsdp.radio.util.JSONParser
 import com.github.fribourgsdp.radio.util.JSONStandardParser
 import junit.framework.TestCase.assertEquals
