+++ app/src/main/AndroidManifest.xml
@@ -46,7 +46,7 @@
             android:name=".game.prep.JoinGameActivity"
             android:exported="false" />
         <activity
-            android:name=".SpotifyReceiveActivity"
+            android:name=".external.spotify.SpotifyReceiveActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:launchMode="singleTop">
@@ -62,7 +62,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".ImportSpotifyPlaylistsActivity"
+            android:name=".external.spotify.ImportSpotifyPlaylistsActivity"
             android:exported="false" />
         <activity
             android:name=".game.prep.LobbyActivity"
